Move captured_output() to common helper module, and removed its local use in test_gamenode and test_historynode
