+++ test/helper.py
@@ -1,0 +1,16 @@
+""" Helper functions for unit tests """
+
+from contextlib import contextmanager
+from io import StringIO
+import sys
+
+@contextmanager
+def captured_output():
+    """ Redirects stdout to StringIO so we can inspect Print statements """
+    new_out = StringIO()
+    old_out = sys.stdout
+    try:
+        sys.stdout = new_out
+        yield sys.stdout
+    finally:
+        sys.stdout = old_out
