+++ test/test_gamenode.py
@@ -1,22 +1,9 @@
 """ Tests for the GameNode module """
 
-from contextlib import contextmanager
-from io import StringIO
-import sys
 import unittest
 
 from src import gamenode
-
-@contextmanager
-def captured_output():
-    """ Redirects stdout to StringIO so we can inspect Print statements """
-    new_out = StringIO()
-    old_out = sys.stdout
-    try:
-        sys.stdout = new_out
-        yield sys.stdout
-    finally:
-        sys.stdout = old_out
+from test import helper
 
 class TestGameNode(unittest.TestCase):
     """ Tests for the GameNode module """
@@ -62,7 +49,7 @@ class TestGameNode(unittest.TestCase):
 
     def test_print_board(self):
         """Check that print_board works"""
-        with captured_output() as out:
+        with helper.captured_output() as out:
             gn_obj = gamenode.GameNode()
             gn_obj.print_board()
             actual_print = out.getvalue().strip()
